timeout prompt flow renamed
